+++ python/pathway/internals/graph_runner/expression_evaluator.py
@@ -448,6 +448,7 @@ class RowwiseEvaluator(
             *(self.eval_expression(arg, eval_state=eval_state) for arg in args),
             propagate_none=expression._propagate_none,
             dtype=expression._dtype.to_engine(),
+            max_batch_size=expression._max_batch_size,
         )
 
     def eval_async_apply(