+++ python/pathway/internals/run.py
@@ -18,6 +18,7 @@ def run(
     persistence_config: PersistenceConfig | None = None,
     runtime_typechecking: bool | None = None,
     terminate_on_error: bool | None = None,
+    max_expression_batch_size: int = 1024,
 ) -> None:
     """Runs the computation graph.
 
@@ -36,6 +37,9 @@ def run(
             persistence is required.
         runtime_typechecking: enables additional strict type checking at runtime
         terminate_on_error: whether to terminate the computation if the data/user-logic error occurs
+        max_expression_batch_size: the maximal number of rows for which the expressions
+            are computed at once. You might want to decrease it if the intermediate state
+            in one of your expressions is large.
     """
     GraphRunner(
         parse_graph.G,
@@ -46,6 +50,7 @@ def run(
         persistence_config=persistence_config,
         runtime_typechecking=runtime_typechecking,
         terminate_on_error=terminate_on_error,
+        max_expression_batch_size=max_expression_batch_size,
         _stacklevel=4,
     ).run_outputs()
 
@@ -60,6 +65,7 @@ def run_all(
     persistence_config: PersistenceConfig | None = None,
     runtime_typechecking: bool | None = None,
     terminate_on_error: bool | None = None,
+    max_expression_batch_size: int = 1024,
 ) -> None:
     """Runs the computation graph with disabled tree-shaking optimization.
 
@@ -78,6 +84,9 @@ def run_all(
             persistence is required.
         runtime_typechecking: enables additional strict type checking at runtime
         terminate_on_error: whether to terminate the computation if the data/user-logic error occurs
+        max_expression_batch_size: the maximal number of rows for which the expressions
+            are computed at once. You might want to decrease it if the intermediate state
+            in one of your expressions is large.
     """
     GraphRunner(
         parse_graph.G,
@@ -88,5 +97,6 @@ def run_all(
         persistence_config=persistence_config,
         runtime_typechecking=runtime_typechecking,
         terminate_on_error=terminate_on_error,
+        max_expression_batch_size=max_expression_batch_size,
         _stacklevel=4,
     ).run_all()